code sniffer fixes
